droidboot: Addition of UEFI compatible recovery_uefi.fstab file
Issue: BMDA-4

The recovery_uefi.fstab file contains the various partition names and
the respective mapping with the device blocks. recovery_uefi.fstab file
in recovery code is required to maintain the signing procedure used to
sign the recovery image. This file is in sync with recovery_uefi.fstab
file used in the droidboot image.

Change-Id: I9fb2f245343bb947126fb7328c278074d0609c03
Signed-off-by: Ravindranath D <ravindranathx.doddi@intel.com>