[REVERT ME] recovery: KK added some mount partition check
BZ: 153083

KitKat recovery checks that only /cache and /tmp partition are
mounted. If any other partitions is mounted, the OTA is aborted.
BUT, we do have another partition mounted and in used (probably)
which /logs. This patch is a temporary workaround to not block
KitKat integration.

Change-Id: I8cd0e31eb0a94026dc0c851ad37437a627ee008a
Category: aosp improvement
Domain: PUPDR-Recovery
Origin: internal
Upstream-Candidate: no
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>