OTA: System enters recovery console twice during OTA
Issue: BMDA 1134

This patch reduces the timeout in recovery mode from 120sec to 60sec.
This is the time we wait in recovery for user to input the
UP + DOWN arrow keys to get the menu. If within this timeout keys are not pressed
we reboot. 120 seconds is a bit long time specially in the OTA scenerio
where system enters in recovery after capsule update. So we reduce that to 60 sec.

Change-Id: I52063812abd84f4469437d87b57ac59b39646811
Signed-off-by: sahashau <shaunak.saha@intel.com>
Reviewed-on: https://android.intel.com/202487
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Deuskar, Prafulla <prafulla.deuskar@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>