increase tmpfs size for sideloaded packages
Recovery Console design specifies that packages provided via
SD Card or 'adb sideload' must be stored in a tmpfs before being
applied. On a 1GB device, the default size of 50% isn't sufficient
for some of the OTA packages we are generating. Increasing to
80% should give us a lot more headroom.

The upstream maintainer of the Recovery Console did not object
to this approach. Forthcoming from upstream is a change to
get rid of this tmpfs buffering for the 'adb sideload' scenario.
Upstream recommended we deprecate SD Card updates.

Issue: GMIN-1993
Change-Id: Idc3534a091fddca5ab3fa1453aeb74ed40d1c38d
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/215192