Fix recovery image build for 32p
When building for 32p, we need to be explicit that we wish to build
the 32bit version of the binaries that will be placed in the recovery
image. The recovery image doesn't actually care... but if we are not
explicit in this, the makefiles will ask for the 64bit binaries but the
Android.mk for the binaries will supply the 32bit images (causing the
build to fail).

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-2414
Change-Id: Iea2d5f412740c082795da4358765751138a4b167
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Reviewed-on: https://android.intel.com:443/260083
Reviewed-on: https://android.intel.com:443/293191