Always use volume mount option when mounting a partition
From ROS, if enable adb through vol.up and down buttons, the
/system partition is mounted by the function
ensure_path_mounted() but with hardcoded mount options.
As a consequence, the blocks are modified and the reboot in MOS
is blocked by the dm_verity feature that detects a corruption.

This patch forces the function ensure_path_mounted() to use
the mount options from the volume structure, that were
previously read from the fstab.

Change-Id: I71ad6493aa3ee51e3ff986c4c74066666587b62f
Category: aosp improvement
Domain: PUPDR-Recovery
Origin: internal
Upstream-Candidate: yes
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-27139
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Reviewed-on: https://android.intel.com:443/348128